feat(cli): bare nika greets, the showroom hands over the keys
One first keystroke, one gesture: bare nika on a TTY = the welcome card (pipes keep the wall + exit 2). nika examples copy <slug> = the adoption step that was missing between the showroom and a founded home; a green Live examples run says so. run/example.rs splits out (file law), main() keeps its cap via the concierge fn.
